Rename DAO package to follow lowercase convetion
