+++ java/src/api/handlers/relators/APIGetRelatedTopicsSubjectHandler.java
@@ -5,7 +5,7 @@ import static api.helpers.RequestMethodHelper.checkRequestMethod;
 import static api.helpers.UrlArgumentHelper.getArgumentsInURL;
 
 import api.exceptions.APIBadRequestException;
-import data.DataAccessObjects.SubjectDAO;
+import data.dao.SubjectDAO;
 import data.Subject;
 import data.Topic;
 import java.util.Collection;
