+++ java/src/data/Topic.java
@@ -1,13 +1,12 @@
 package data;
 
-import data.DataAccessObjects.TopicDAO;
+import data.dao.TopicDAO;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
-import javax.persistence.PersistenceContext;
 import javax.persistence.Table;
