+++ java/src/data/dao/SubjectDAO.java
@@ -1,6 +1,6 @@
-package data.DataAccessObjects;
+package data.dao;
 
-import data.DataAccessObjects.util.FieldTuple;
+import data.dao.util.FieldTuple;
 import data.Subject;
 import data.Topic;
 import data.user.User;
