+++ java/tests/api/handlers/user/APILoginHandlerTest.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.*;
 import api.exceptions.APIBadMethodException;
 import api.exceptions.APIBadRequestException;
 import base.BaseTest;
-import data.DataAccessObjects.UserDAO;
+import data.dao.UserDAO;
 import data.user.User;
 import java.io.ByteArrayInputStream;
 import org.apache.http.HttpRequest;
